+++ packages/core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log - @sap/guided-answers-extension-core
 
+## 0.14.0
+
+### Minor Changes
+
+-   38594b7: feat: #67 - blogs search result page is sorted by relevance or time updated
+
+### Patch Changes
+
+-   Updated dependencies [38594b7]
+    -   @sap/knowledge-hub-extension-types@0.14.0
+
 ## 0.13.0
 
 ### Patch Changes